requirements.txt
numpy==1.23.4
pandas==1.5.1
matplotlib==3.6.1 
scipy==1.10.0
pyzmq==24.0.1 
requests==2.28.2
python-binance==1.0.16
binance-connector
websocket-client